chore(cli): extract trace CLI command handlers into separate files
