+++ tests/mcp/trace-cli.spec.ts
@@ -20,8 +20,8 @@ import { test, expect } from './trace-cli-fixtures';
 
 test.skip(({ mcpBrowser }) => mcpBrowser !== 'chrome', 'Chrome-only');
 
-test('trace info shows metadata', async ({ traceFile, runTraceCli }) => {
-  const { stdout, exitCode } = await runTraceCli(['info', traceFile]);
+test('trace open shows metadata', async ({ traceFile, runTraceCli }) => {
+  const { stdout, exitCode } = await runTraceCli(['open', traceFile]);
   expect(exitCode).toBe(0);
   expect(stdout).toContain('Browser:');
   expect(stdout).toContain('chromium');
@@ -32,8 +32,8 @@ test('trace info shows metadata', async ({ traceFile, runTraceCli }) => {
   expect(stdout).toContain('Network:');
 });
 
-test('trace actions shows actions with ordinals', async ({ traceFile, runTraceCli }) => {
-  const { stdout, exitCode } = await runTraceCli(['actions', traceFile]);
+test('trace actions shows actions with ordinals', async ({ runTraceCli }) => {
+  const { stdout, exitCode } = await runTraceCli(['actions']);
   expect(exitCode).toBe(0);
   // Should have ordinal numbers
   expect(stdout).toMatch(/^\s+\d+\.\s/m);
@@ -43,21 +43,21 @@ test('trace actions shows actions with ordinals', async ({ traceFile, runTraceCl
   expect(stdout).toContain('Fill');
 });
 
-test('trace actions --grep filters actions', async ({ traceFile, runTraceCli }) => {
-  const { stdout, exitCode } = await runTraceCli(['actions', '--grep', 'Click', traceFile]);
+test('trace actions --grep filters actions', async ({ runTraceCli }) => {
+  const { stdout, exitCode } = await runTraceCli(['actions', '--grep', 'Click']);
   expect(exitCode).toBe(0);
   expect(stdout).toContain('Click');
   expect(stdout).not.toContain('Navigate');
   expect(stdout).not.toContain('Fill');
 });
 
-test('trace action displays action details', async ({ traceFile, runTraceCli }) => {
+test('trace action displays action details', async ({ runTraceCli }) => {
   // First get an action ordinal from list
-  const { stdout: listOutput } = await runTraceCli(['actions', '--grep', 'Navigate', traceFile]);
+  const { stdout: listOutput } = await runTraceCli(['actions', '--grep', 'Navigate']);
   const match = listOutput.match(/^\s+(\d+)\.\s/m);
   expect(match).toBeTruthy();
 
-  const { stdout, exitCode } = await runTraceCli(['action', traceFile, match![1]]);
+  const { stdout, exitCode } = await runTraceCli(['action', match![1]]);
   expect(exitCode).toBe(0);
   expect(stdout).toContain('Navigate');
   expect(stdout).toContain('Time');
@@ -66,25 +66,25 @@ test('trace action displays action details', async ({ traceFile, runTraceCli }) 
   expect(stdout).toContain('Parameters');
 });
 
-test('trace action reports available snapshots', async ({ traceFile, runTraceCli }) => {
-  const { stdout: listOutput } = await runTraceCli(['actions', '--grep', 'Click', traceFile]);
+test('trace action reports available snapshots', async ({ runTraceCli }) => {
+  const { stdout: listOutput } = await runTraceCli(['actions', '--grep', 'Click']);
   const match = listOutput.match(/^\s+(\d+)\.\s/m);
   expect(match).toBeTruthy();
 
-  const { stdout, exitCode } = await runTraceCli(['action', traceFile, match![1]]);
+  const { stdout, exitCode } = await runTraceCli(['action', match![1]]);
   expect(exitCode).toBe(0);
   expect(stdout).toContain('Snapshots');
   expect(stdout).toContain('available:');
 });
 
-test('trace action with invalid action ID', async ({ traceFile, runTraceCli }) => {
-  const { stderr, exitCode } = await runTraceCli(['action', traceFile, '999999']);
+test('trace action with invalid action ID', async ({ runTraceCli }) => {
+  const { stderr, exitCode } = await runTraceCli(['action', '999999']);
   expect(exitCode).toBe(1);
   expect(stderr).toContain('not found');
 });
 
-test('trace requests shows requests with ordinals', async ({ traceFile, runTraceCli }) => {
-  const { stdout, exitCode } = await runTraceCli(['requests', traceFile]);
+test('trace requests shows requests with ordinals', async ({ runTraceCli }) => {
+  const { stdout, exitCode } = await runTraceCli(['requests']);
   expect(exitCode).toBe(0);
   expect(stdout).toContain('Method');
   expect(stdout).toContain('Status');
@@ -92,15 +92,15 @@ test('trace requests shows requests with ordinals', async ({ traceFile, runTrace
   expect(stdout).toMatch(/\d+\./);
 });
 
-test('trace requests --method filters', async ({ traceFile, runTraceCli }) => {
-  const { stdout, exitCode } = await runTraceCli(['requests', '--method', 'GET', traceFile]);
+test('trace requests --method filters', async ({ runTraceCli }) => {
+  const { stdout, exitCode } = await runTraceCli(['requests', '--method', 'GET']);
   expect(exitCode).toBe(0);
   expect(stdout).toContain('GET');
   expect(stdout).not.toContain('POST');
 });
 
-test('trace request shows details', async ({ traceFile, runTraceCli }) => {
-  const { stdout, exitCode } = await runTraceCli(['request', traceFile, '1']);
+test('trace request shows details', async ({ runTraceCli }) => {
+  const { stdout, exitCode } = await runTraceCli(['request', '1']);
   expect(exitCode).toBe(0);
   expect(stdout).toContain('General');
   expect(stdout).toContain('status:');
@@ -108,66 +108,61 @@ test('trace request shows details', async ({ traceFile, runTraceCli }) => {
   expect(stdout).toContain('Response headers');
 });
 
-test('trace request with invalid ID', async ({ traceFile, runTraceCli }) => {
-  const { stderr, exitCode } = await runTraceCli(['request', traceFile, '999999']);
+test('trace request with invalid ID', async ({ runTraceCli }) => {
+  const { stderr, exitCode } = await runTraceCli(['request', '999999']);
   expect(exitCode).toBe(1);
   expect(stderr).toContain('not found');
 });
 
-test('trace console shows messages', async ({ traceFile, runTraceCli }) => {
-  const { stdout, exitCode } = await runTraceCli(['console', traceFile]);
+test('trace console shows messages', async ({ runTraceCli }) => {
+  const { stdout, exitCode } = await runTraceCli(['console']);
   expect(exitCode).toBe(0);
   expect(stdout).toContain('info message');
   expect(stdout).toContain('warning message');
   expect(stdout).toContain('error message');
 });
 
-test('trace console --errors-only', async ({ traceFile, runTraceCli }) => {
-  const { stdout, exitCode } = await runTraceCli(['console', '--errors-only', traceFile]);
+test('trace console --errors-only', async ({ runTraceCli }) => {
+  const { stdout, exitCode } = await runTraceCli(['console', '--errors-only']);
   expect(exitCode).toBe(0);
   expect(stdout).toContain('error message');
   expect(stdout).not.toContain('info message');
 });
 
-test('trace errors', async ({ traceFile, runTraceCli }) => {
-  const { stdout, exitCode } = await runTraceCli(['errors', traceFile]);
+test('trace errors', async ({ runTraceCli }) => {
+  const { stdout, exitCode } = await runTraceCli(['errors']);
   expect(exitCode).toBe(0);
   // Our test trace may or may not have errors, just verify it doesn't crash
   expect(stdout).toBeTruthy();
 });
 
-test('trace snapshot saves HTML file', async ({ traceFile, runTraceCli }, testInfo) => {
-  const { stdout: listOutput } = await runTraceCli(['actions', '--grep', 'Navigate', traceFile]);
+test('trace snapshot runs command on snapshot', async ({ runTraceCli }) => {
+  const { stdout: listOutput } = await runTraceCli(['actions', '--grep', 'Navigate']);
   const match = listOutput.match(/^\s+(\d+)\.\s/m);
   expect(match).toBeTruthy();
 
-  const outPath = testInfo.outputPath('test-snapshot.html');
-  const { stdout, exitCode } = await runTraceCli(['snapshot', traceFile, match![1], '-o', outPath]);
+  const { stdout, exitCode } = await runTraceCli(['snapshot', match![1]]);
   expect(exitCode).toBe(0);
-  expect(stdout).toContain('Snapshot saved to');
-  expect(fs.existsSync(outPath)).toBe(true);
-  const html = fs.readFileSync(outPath, 'utf-8');
-  expect(html.toLowerCase()).toContain('<html');
+  expect(stdout).toBeTruthy();
 });
 
-test('trace snapshot --name before', async ({ traceFile, runTraceCli }, testInfo) => {
-  const { stdout: listOutput } = await runTraceCli(['actions', '--grep', 'Click', traceFile]);
+test('trace snapshot --name before', async ({ runTraceCli }) => {
+  const { stdout: listOutput } = await runTraceCli(['actions', '--grep', 'Click']);
   const match = listOutput.match(/^\s+(\d+)\.\s/m);
   expect(match).toBeTruthy();
 
-  const outPath = testInfo.outputPath('before-snapshot.html');
-  const { stdout, exitCode } = await runTraceCli(['snapshot', '--name', 'before', traceFile, match![1], '-o', outPath]);
+  const { stdout, exitCode } = await runTraceCli(['snapshot', '--name', 'before', match![1]]);
   expect(exitCode).toBe(0);
-  expect(stdout).toContain('Snapshot saved to');
+  expect(stdout).toBeTruthy();
 });
 
-test('trace screenshot saves image file', async ({ traceFile, runTraceCli }, testInfo) => {
-  const { stdout: listOutput } = await runTraceCli(['actions', '--grep', 'Navigate', traceFile]);
+test('trace screenshot saves image file', async ({ runTraceCli }, testInfo) => {
+  const { stdout: listOutput } = await runTraceCli(['actions', '--grep', 'Navigate']);
   const match = listOutput.match(/^\s+(\d+)\.\s/m);
   expect(match).toBeTruthy();
 
   const outPath = testInfo.outputPath('test-screenshot.png');
-  const { stdout, exitCode } = await runTraceCli(['screenshot', traceFile, match![1], '-o', outPath]);
+  const { stdout, exitCode } = await runTraceCli(['screenshot', match![1], '-o', outPath]);
   // Screenshot may or may not be available depending on timing
   if (exitCode === 0) {
     expect(stdout).toContain('Screenshot saved to');
@@ -175,8 +170,8 @@ test('trace screenshot saves image file', async ({ traceFile, runTraceCli }, tes
   }
 });
 
-test('trace attachments lists attachments', async ({ traceFile, runTraceCli }) => {
-  const { stdout, exitCode } = await runTraceCli(['attachments', traceFile]);
+test('trace attachments lists attachments', async ({ runTraceCli }) => {
+  const { stdout, exitCode } = await runTraceCli(['attachments']);
   expect(exitCode).toBe(0);
   // Our test trace has no attachments, just verify it doesn't crash
   expect(stdout).toBeTruthy();
